Renames files.
As part of this re-org, I am putting the old install-this and that into
a folder called kubuntu. My new, current stuff, will go into the ubuntu
folder. If I decide to stick with Ubuntu, I may delete the kubuntu files
and move the ubuntu files to the root folder.

 Changes to be committed:
	new file:	investigate.txt
	renamed:	install-art.sh -> kubuntu/install-art.sh
	renamed:	install-desktop.sh -> kubuntu/install-desktop.sh
	renamed:	install-notes.md -> kubuntu/install-notes.md
	renamed:	install-programming.sh -> kubuntu/install-programming.sh